Imports data on House of Lords Amendments. Returns a tibble with all available House of Lords amendments.

Imports data on House of Lords Amendments. Returns a tibble with all available House of Lords amendments.

lords_amendments(decision = NULL, start_date = "1900-01-01",
  end_date = Sys.Date(), extra_args = NULL, tidy = TRUE)

Arguments

decision
The decision on the amendments. Accepts one of 'Withdrawn', 'Agreed', 'Disagreed', 'Pending', 'NotMoved', 'Disposed'. Defaults to NULL.

start_date
The earliest date to include in the tibble. Defaults to '1900-01-01'.

end_date
The latest date to include in the tibble. Defaults to current system date.

extra_args
Additional parameters to pass to API. Defaults to NULL.

tidy
Fix the variable names in the tibble to remove extra characters, superfluous text and convert variable names to snake_case. Defaults to TRUE.

Value

A tibble with details on amendments proposed by the House of Lords.

Examples

not_run({

  x <- lords_amendments()

})