Members of both houses

Imports basic details on current and former Members of Parliament including the Lords and the Commons.

members(ID = NULL, extra_args = NULL, tidy = TRUE,
  tidy_style = "snake_case", verbose = FALSE)

hansard_members(ID = NULL, extra_args = NULL, tidy = TRUE,
  tidy_style = "snake_case", verbose = FALSE)

commons_members(extra_args = NULL, tidy = TRUE, tidy_style = "snake_case",
  verbose = FALSE)

hansard_commons_members(extra_args = NULL, tidy = TRUE,
  tidy_style = "snake_case", verbose = FALSE)

lords_members(extra_args = NULL, tidy = TRUE, tidy_style = "snake_case",
  verbose = FALSE)

hansard_lords_members(extra_args = NULL, tidy = TRUE,
  tidy_style = "snake_case", verbose = FALSE)

Arguments

ID
The ID of a member of the House of Commons or the House of Lords to return data on. If NULL, returns a tibble of all members of both houses. Defaults to NULL.

extra_args
Additional parameters to pass to API. See the API documentation or the package vignette for more details. Defaults to NULL.

tidy
Logical parameter. If TRUE, fixes the variable names in the tibble to remove special characters and superfluous text, and converts the variable names to a consistent style. Defaults to TRUE.

tidy_style
The style to convert variable names to, if tidy = TRUE. Accepts one of 'snake_case', 'camelCase' and 'period.case'. Defaults to 'snake_case'.

verbose
If TRUE, returns data to console on the progress of the API request. Defaults to FALSE.

Value

A tibble with data on members of the House of Commons (commons_members()), the House of Lords, (lords_members()), or both (members()).

Details

For more details on a given member see mnis_full_biog.

Member details functions

members

Basic details on a given member from either house

commons_members

MPs in the House of Commons

lords_members

Peers in the House of Lords

See also

members_search

Examples

# NOT RUN {
a <- members()

x <- members(172)

y <- commons_members()

z <- lords_members()
# }